% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopVisorPos.R
\name{TopVisorPos}
\alias{TopVisorPos}
\title{Get positions}
\usage{
TopVisorPos(user_id = NULL, token = NULL, project_id = NULL,
  date1 = NULL, date2 = NULL)
}
\arguments{
\item{user_id}{Your user ID, see \code{\link{https://topvisor.com/account/}}}

\item{token}{Your API token, see \code{\link{https://topvisor.com/account/}}}

\item{project_id}{Project ID}

\item{date1}{Date from}

\item{date2}{Date till}
}
\description{
Function to get positions
}
\examples{
TopVisorPos()
}
